refactor: OrderTable 패키지 변경
