+++ src/main/java/kitchenpos/repository/order/OrderTableRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.repository.order;
 
-import kitchenpos.domain.order.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
