+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -5,7 +5,7 @@ import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItem;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.order.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 import kitchenpos.dto.order.request.OrderLineItemRequest;
 import kitchenpos.dto.order.request.OrderRequest;
 import kitchenpos.dto.order.response.OrderResponse;
