+++ src/test/java/kitchenpos/domain/OrderTableTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain;
 
-import kitchenpos.domain.order.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.table.TableGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
