+++ src/test/java/kitchenpos/fixture/OrderTableFixture.java
@@ -1,6 +1,6 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.order.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.table.TableGroup;
 
 import static kitchenpos.fixture.TableGroupFixture.createTableGroupWithoutId;
