+++ src/test/java/kitchenpos/ui/TableRestControllerTest.java
@@ -3,7 +3,7 @@ package kitchenpos.ui;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import kitchenpos.application.table.TableService;
-import kitchenpos.domain.order.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 import kitchenpos.dto.table.response.OrderTableResponse;
 import kitchenpos.fixture.OrderTableFixture;
 import kitchenpos.ui.table.TableRestController;
